Matlab programs
##  Classifier Neural Network
Make a simple classifier neural network using the delta rule. 
This model has four input nodes and two output nodes. It must learn how to map four particular input patterns to four particular output patterns. The delta rule allows it to compare its actual output on any given trial with the target (or correct) output, and use that difference (error signal) to adjust the weights appropriately. 

## Two Layer Back-propagation network
Make a two-layer backpropagation network that solves the XOR problem. 

## Logistic Map
The Logistic Map (y=h*y*(1-y)) Conduct 4000 runs of the logistic map going through 1000 timesteps each, with each run using a different h parameter value from 0 to 4 in steps of .001 From each run, save the last 40 values of y (which will be kept track of with variable x). And then plot them all to show the classic bifurcation plot of the logistic map. 

## Mandlebrot
 Plot the FORMAL (black-and-white) Mandelbrot Set, using this iterative equation: z = z2 + c where z starts at zero, and c is a complex number that ranges from -2 to 1 in the real component and -1 to 1 in the imaginary component. If that iterative process does not escape to infinity, then that complex number belongs to the Mandelbrot set. Note that rather than testing whether any given iteration really escapes to infinity (and therefore is not formally a part of the Mandelbrot set), you can actually just check whether the number goes above 2. (That is, if the value gets past 2, then the iterative process WILL escape to infinity.) 

## Lorenz Attractor
Plot the Lorenz Attractor 
The Lorenz system is a system of ordinary differential equations first studied by Edward Lorenz. It is notable for having chaotic solutions for certain parameter values and initial conditions. In particular, the Lorenz attractor is a set of chaotic solutions of the Lorenz system which, when plotted, resemble a butterfly or figure eight.

## Game Of Life
Reproduce the Game Of Life. 
The "game" is a zero-player game, meaning that its evolution is determined by its initial state, requiring no further input. One interacts with the Game of Life by creating an initial configuration and observing how it evolves, or, for advanced "players", by creating patterns with particular properties.